fix: type int to long (#100)

+++ core/src/main/kotlin/com/retoday/core/domain/recap/dto/projection/UserActivityProjection.kt
@@ -5,5 +5,5 @@ data class UserActivityProjection(
     val description: String?,
     val domain: String,
     val categoryName: String?,
-    val stayDuration: Int
+    val stayDuration: Long
 )
